Receive SIP identifiers
